==> Code/1_reading_dataset.ipynb <==
### Import libraries

import pandas as pd
import re
from nltk.corpus import stopwords
import nltk
import warnings
warnings.filterwarnings('ignore')
import logging
logging.getLogger("lda").setLevel(logging.WARNING)
try:
  nltk.download('stopwords')
except:
  pass
try:
  nltk.download('punkt')
except:
  pass

##### Import stopwords

stopwords = stopwords.words('english')

## Reading Questions.csv

df_questions = pd.read_csv("Questions.csv",usecols=['Id', 'CreationDate', 'Score', 'Title','Body'], encoding='latin1')
df_questions = df_questions.dropna()
df_questions.head()

## Reading Answers.csv

df_answers = pd.read_csv("Answers.csv",usecols = ['ParentId','Score','Body'], encoding='latin1')
df_answers.columns = ['Id','Answer_Score','Answer_Body']
df_answers

## Reading Tags.csv

df_tags = pd.read_csv("Tags.csv" ,encoding='latin1')
df_tags

## Preprocessing Datasets

### 1. Define functions

#####  Normalize Text: normalize_text() to remove tags of html and codes etc

def normalize_text(text):
    tm1 = re.sub('<pre>.*?</pre>', '', text, flags=re.DOTALL)
    tm2 = re.sub('<code>.*?</code>', '', tm1, flags=re.DOTALL)
    tm3 = re.sub('<[^>]+>', '', tm2, flags=re.DOTALL)
    x = tm3.replace("\n", "")
    return x

#### Remove Stopwords and Punctuations: remove_stopwords_punctuations() to remove them

def remove_stopwords_punctuations(text):
    #REmove punctuations ~30 seconds
    text = text.replace(r'[^\w\s]+', '')
    # remove stopwords ~4 minutes
    text = ' '.join([word for word in text.split() if word not in (stopwords)])
    return text    

#### Concatenate multiple tags: same question having multiple tags is handled

def concatenate_multiple_tags(df):
    df = df.groupby('Id', as_index=False).agg({'Tag' : ','.join})
    return df

### 2. Application of preprocessing functions on datasets

#### on Questions.csv

df_questions['Body'] = df_questions['Body'].apply(lambda x: normalize_text(x))
df_questions['Body'] = df_questions['Body'].apply(lambda x: remove_stopwords_punctuations(x))
df_questions

#### on Answers.csv

df_answers['Answer_Body'] = df_answers['Answer_Body'].apply(lambda x: normalize_text(x))
df_answers['Answer_Body'] = df_answers['Answer_Body'].apply(lambda x: remove_stopwords_punctuations(x))
df_answers

#### on Tags.csv

df_tags['Tag'] = df_tags['Tag'].apply(lambda x: str(x))
df_tags = concatenate_multiple_tags(df_tags)
df_tags

#### Merging dataframes

merged = pd.merge(df_questions, df_answers, how='inner', on="Id")

merged = pd.merge(merged, df_tags, how='inner', on="Id")

merged

merged = merged.astype("string")
merged[['Body', 'Tag']].fillna("", inplace = True)
# merged

merged.to_csv("questions_answers_tags.csv", index = False, encoding = 'latin1')
