==> Code/2_reduce_dataset.ipynb <==
### Import libraries

import pandas as pd

## Reading questions_answers_tags.csv

merged = pd.read_csv("questions_answers_tags.csv", encoding = 'latin1')

## Remove Answer_Body and Answer_Score columns

merged = merged.drop(['Answer_Score','Answer_Body'],axis=1)

## Drop duplicates that have been created due to removal of Answer data

merged = merged.drop_duplicates()

## Saving QAs_cleaned.csv file

merged.to_csv('QAs_cleaned.csv',encoding='latin1')
