==> Code/3_fse.ipynb <==
import pandas as pd
import numpy as np

!pip install fse


from fse import Vectors, Average, IndexedList
vecs = Vectors.from_pretrained("glove-wiki-gigaword-300")
model = Average(vecs)

sentences = [["cat", "say", "meow"], ["dog", "say", "woof"]]

df = pd.read_csv("QAs_cleaned.csv",usecols = ['Id','CreationDate','Score', 'Title', 'Body','Tag'],encoding='latin1')
df = df.drop_duplicates()
df.reset_index(drop=True, inplace=True)
df

df['title_vec'] = df['Title']+df['Tag']

df.title_vec

df['title_vec'] = df['title_vec'].apply(lambda x: model.infer([(str(x).split(), 0)])[0].tolist())

df['title_vec'][0]

df.to_pickle("final1.pkl")

df.dtypes
