==> Code/4_ingest.ipynb <==
import pandas as pd
import warnings
warnings.simplefilter("ignore")
import json
import pandas as pd
import requests
import logging
logging.getLogger('requests').setLevel(logging.CRITICAL)

from elasticsearch import Elasticsearch
es = Elasticsearch(hosts = [
    'http://localhost:9201'
])

es.info()

import numpy as np
import pandas as pd

df = pd.read_pickle("final1.pkl")

df

try:
    es.indices.create(index = 'so-vec-ret')
except:
    pass

df.columns

df.dtypes

from json import JSONEncoder


class NumpyArrayEncoder(JSONEncoder):
    def default(self, obj):
        if isinstance(obj, np.ndarray):
            return obj.tolist()
        return JSONEncoder.default(self, obj)

df_split = np.array_split(df, 4)

index = 'so-vec-ret'
typ = 'test'
server = 'http://localhost:9201'

headers = {'content-type': 'application/x-ndjson', 'Accept-Charset': 'UTF-8'}
records = df_split[0].to_dict(orient='records')
actions = ["""{ "index" : { "_index" : "%s", "_type" : "%s"} }\n""" % (index, typ) +json.dumps(records[j])
                for j in range(len(records))]

chunk_size = 5000

#actions1 = [json.dumps(records[j], cls=NumpyArrayEncoder) for j in ]
        
i=0
while i<len(actions):
    if(i%10000 == 0):
        print(i)
    try:
        serverAPI = server + '/_bulk' 
        data='\n'.join(actions[i:min([i+chunk_size,len(actions)])])
        data = data + '\n'
        r = requests.post(serverAPI, data = data, headers=headers)
        #print (r.content)
    except:
        break
    i = i+chunk_size

index = 'so-vec-ret'
typ = 'test'
server = 'http://localhost:9201'

headers = {'content-type': 'application/x-ndjson', 'Accept-Charset': 'UTF-8'}
records = df_split[1].to_dict(orient='records')
actions = ["""{ "index" : { "_index" : "%s", "_type" : "%s"} }\n""" % (index, typ) +json.dumps(records[j])
                for j in range(len(records))]

chunk_size = 5000

#actions1 = [json.dumps(records[j], cls=NumpyArrayEncoder) for j in ]
        
i=0
while i<len(actions):
    if(i%10000 == 0):
        print(i)
    try:
        serverAPI = server + '/_bulk' 
        data='\n'.join(actions[i:min([i+chunk_size,len(actions)])])
        data = data + '\n'
        r = requests.post(serverAPI, data = data, headers=headers)
        #print (r.content)
    except:
        break
    i = i+chunk_size

index = 'so-vec-ret'
typ = 'test'
server = 'http://localhost:9201'

headers = {'content-type': 'application/x-ndjson', 'Accept-Charset': 'UTF-8'}
records = df_split[2].to_dict(orient='records')
actions = ["""{ "index" : { "_index" : "%s", "_type" : "%s"} }\n""" % (index, typ) +json.dumps(records[j])
                for j in range(len(records))]
chunk_size = 5000

#actions1 = [json.dumps(records[j], cls=NumpyArrayEncoder) for j in ]
        
i=0
while i<len(actions):
    if(i%10000 == 0):
        print(i)
    try:
        serverAPI = server + '/_bulk' 
        data='\n'.join(actions[i:min([i+chunk_size,len(actions)])])
        data = data + '\n'
        r = requests.post(serverAPI, data = data, headers=headers)
        #print (r.content)
    except:
        break
    i = i+chunk_size

index = 'so-vec-ret'
typ = 'test'
server = 'http://localhost:9201'

headers = {'content-type': 'application/x-ndjson', 'Accept-Charset': 'UTF-8'}
records = df_split[3].to_dict(orient='records')
actions = ["""{ "index" : { "_index" : "%s", "_type" : "%s"} }\n""" % (index, typ) +json.dumps(records[j])
                for j in range(len(records))]

chunk_size = 5000

#actions1 = [json.dumps(records[j], cls=NumpyArrayEncoder) for j in ]
        
i=0
while i<len(actions):
    if(i%10000 == 0):
        print(i)
    try:
        serverAPI = server + '/_bulk' 
        data='\n'.join(actions[i:min([i+chunk_size,len(actions)])])
        data = data + '\n'
        r = requests.post(serverAPI, data = data, headers=headers)
        #print (r.content)
    except:
        break
    i = i+chunk_size

try:
    b = {
        "mappings": {
            "properties": {
                "Title": {
                    "type": "text"
                },
                "title_vec": {
                    "type": "dense_vector",
                    "dims": 300
                }
            }
        }
    }


    ret = es.indices.create(index='so-vec-ret', ignore=400, body=b) #400 caused by IndexAlreadyExistsException, 
    print(json.dumps(ret,indent=4))
except:
    print("cant")
